Collapse (es) and en/ route groups into unified canonical paths
All routes now serve the same path regardless of language — no /en/ prefix.
Spanish slugs are canonical (/, /aniversario, /codigo-conducta, /terminos,
/patrocinadores, /aniversario/patrocinadores).

- Promote (es)/layout.tsx to root layout, wire LangProvider
- Promote all (es) page files to root routes
- Remove src/app/en/ entirely (7 files: layout, pages, sub-routes)
- Remove src/app/(es)/ entirely (route group no longer needed)
