--- src/app/en/sponsors/page.tsx
@@ -1,1 +1,0 @@
-export { metadata, default } from "../aniversario/sponsors/page";
