+++ src/app/terminos/page.tsx
